feat: add a button to copy pull URL
